Refactoring: move distances into knn_utils.rs
